requirements.txt
dateparser==1.0.0
selenium==3.141.0
requests==2.26.0
numpy==1.20.3
aiogram==2.14.3
python-dotenv==0.19.0
